# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed:
    properties:
      error:
        example: Movie not found in job
        type: string
      movie_id:
        example: IPX-535
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobListResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
        type: array
    type: object
...
  internal_api_batch.BatchExcludeRequest:
    properties:
      movie_ids:
        example:
        - IPX-535
        - ABC-123
        items:
          type: string
        type: array
    required:
    - movie_ids
    type: object
  internal_api_batch.BatchExcludeResponse:
    properties:
      excluded:
        items:
          type: string
        type: array
      failed:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed'
        type: array
      job:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
    type: object
  internal_api_batch.BatchJobResponse:
    properties:
      completed:
        type: integer
      completed_at:
        type: string
      destination:
        type: string
      excluded:
        additionalProperties:
          type: boolean
        type: object
      failed:
        type: integer
      id:
        type: string
      operation_count:
        type: integer
      operation_mode_override:
        type: string
      persist_error:
        type: string
      progress:
        type: number
      results:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult'
        type: object
      reverted_count:
        type: integer
      started_at:
        type: string
      status:
        type: string
      total_files:
        type: integer
      update:
        type: boolean
    type: object
...
  internal_api_batch.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
  internal_api_batch.MovieResponse:
    properties:
      merge_stats:
        allOf:
        - $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics'
        description: Merge statistics when NFO merging occurred
      movie:
        $ref: '#/definitions/models.Movie'
      provenance:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DataSource'
        description: Field-level data source tracking
        type: object
    type: object
...
  internal_api_batch.PosterCropRequest:
    properties:
      height:
        minimum: 1
        type: integer
      width:
        minimum: 1
        type: integer
      x:
        minimum: 0
        type: integer
      "y":
        minimum: 0
        type: integer
    type: object
  internal_api_batch.PosterCropResponse:
    properties:
      cropped_poster_url:
        type: string
    type: object
  internal_api_batch.PosterFromURLRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  internal_api_batch.PosterFromURLResponse:
    properties:
      cropped_poster_url:
        type: string
      poster_url:
        type: string
    type: object
  internal_api_batch.UpdateMovieRequest:
    properties:
      movie:
        $ref: '#/definitions/models.Movie'
    required:
    - movie
    type: object
...
  internal_api_movie.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  internal_api_movie.MoviesResponse:
    properties:
      count:
        example: 20
        type: integer
      movies:
        items:
          $ref: '#/definitions/models.Movie'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/fedora-oss/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/fedora-oss/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/batch:
    get:
      description: Get a list of batch jobs with operation counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: List batch jobs
      tags:
      - web
  /api/v1/batch/{id}:
    delete:
      description: Delete a completed or cancelled batch job and its temp files. Running
        jobs must be cancelled first.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Delete batch job
      tags:
      - web
    get:
      description: Retrieve the status of a batch scraping job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_batch.BatchJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Get batch job status
      tags:
      - web
...
  /api/v1/batch/{id}/movies/{movieId}:
    patch:
      consumes:
      - application/json
      description: Update a movie's metadata within a batch job's results
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      - description: Updated movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_batch.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_batch.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Update movie in batch job
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}/exclude:
    post:
      description: Mark a movie in a batch job as excluded from file organization
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Exclude movie from batch organization
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}/poster-crop:
    post:
      consumes:
      - application/json
      description: Re-crop a temp poster for the review page using fixed-size crop
        coordinates
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      - description: Crop coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_batch.PosterCropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_batch.PosterCropResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Update manual poster crop in batch job
      tags:
      - web
  /api/v1/batch/{id}/movies/{movieId}/poster-from-url:
    post:
      consumes:
      - application/json
      description: Download a poster image from a URL and set it as the movie's poster
        in the batch job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: string
      - description: Poster URL
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_batch.PosterFromURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_batch.PosterFromURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Download poster from URL
      tags:
      - web
...
  /api/v1/batch/{id}/movies/batch-exclude:
    post:
      consumes:
      - application/json
      description: 'Exclude multiple movies from a batch job in a single request.
        Best-effort: excludes as many as possible and returns per-movie failures.'
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Movie IDs to exclude
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_batch.BatchExcludeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_batch.BatchExcludeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_batch.ErrorResponse'
      summary: Bulk exclude movies from batch organization
      tags:
      - web
...
  /api/v1/movies:
    get:
      description: Get a paginated list of all movies cached in the database. Supports
        pagination via limit and offset query parameters. Returns movie count and
        basic metadata.
      parameters:
      - description: Max number of movies to return
        in: query
        name: limit
        type: integer
      - description: Number of movies to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_movie.MoviesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_movie.ErrorResponse'
      summary: List cached movies
      tags:
      - movies
  /api/v1/movies/{id}:
    get:
      description: Retrieve movie metadata from cache by ID, optionally with provenance
        information
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      - description: Include field-level provenance data
        in: query
        name: include_provenance
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_movie.MovieResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_movie.ErrorResponse'
      summary: Get movie by ID
      tags:
      - movies
...
schemes:
- http
- https
swagger: "2.0"